[1.0.2] Various Bugs Fixed
Fixed:

- Fixed an issue from URP where motion vectors in scene view were incorrectly rendered in Unity 2022 & 2023.
- Fixed an issue where previous depth texture was incorrectly set as the current one, causing severe ghosting.
- Fixed an issue where textures may not be set when Render Graph is enabled.
- Fixed an issue with Automatic Thickness Mode.
- Fixed an issue with Indirect Diffuse Rendering Layers.
- Fixed an issue with incorrect surface normals when Accurate GBuffer Normals was enabled.
- Fixed an issue with random sampling direction being broken on platforms using half-precision floats.

+++ Shaders/SSGIInput.hlsl
@@ -110,7 +110,7 @@ half4 ssgi_SHBb;
 half4 ssgi_SHC;
 
 half _DownSample;
-half _FrameIndex;
+float _FrameIndex;
 half _MaxMediumSteps;
 half _MaxSmallSteps;
 half _HistoryTextureValid;
